requirements.txt
# Core Dependencies
appium-python-client==2.11.1
selenium==4.15.2

# Android Automation
pure-python-adb==0.3.0.dev0
uiautomator2==3.4.2

# Image Processing
Pillow==10.1.0
opencv-python==4.8.1.78

# Report Generation
jinja2==3.1.2
markdown==3.5.1
weasyprint==60.1

# Utilities
pyyaml==6.0.1
requests==2.31.0
colorama==0.4.6
tqdm==4.66.1

# Machine Learning (Optional for advanced features)
numpy==1.24.3
scikit-learn==1.3.2

# Logging
loguru==0.7.2
